1088: Removing all OB Funds Confirmations API versions older than v3.1.10
Older API versions are not supported by the Consent Store, and are so old that no ASPSP should be offering them.

Where v3.1.10 controllers are extending older version controllers, the logic from the version with the impl being used has been pulled up into the v3.1.10 controller.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1088
