--- secure-api-gateway-ob-uk-rs-obie-api/src/main/java/com/forgerock/sapi/gateway/ob/uk/rs/obie/api/funds/v3_0/FundsConfirmationsApi.java
@@ -1,88 +1,0 @@
-/*
- * Copyright © 2020-2024 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.forgerock.sapi.gateway.ob.uk.rs.obie.api.funds.v3_0;
-
-import com.forgerock.sapi.gateway.ob.uk.common.error.OBErrorResponseException;
-import com.forgerock.sapi.gateway.ob.uk.rs.obie.api.ApiConstants;
-import com.forgerock.sapi.gateway.ob.uk.rs.obie.api.swagger.SwaggerApiTags;
-import io.swagger.annotations.*;
-import org.joda.time.DateTime;
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import uk.org.openbanking.datamodel.error.OBErrorResponse1;
-import uk.org.openbanking.datamodel.fund.OBFundsConfirmation1;
-import uk.org.openbanking.datamodel.fund.OBFundsConfirmationResponse1;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.validation.Valid;
-import java.security.Principal;
-
-@Api(tags = {"v3.0", SwaggerApiTags.CONFIRMATION_OF_FUNDS_TAG})
-@RequestMapping(value = "/open-banking/v3.0/cbpii")
-public interface FundsConfirmationsApi {
-
-    String FUNDS_CONFIRMATION_PATH = "/funds-confirmations";
-
-    @ApiOperation(value = "Create Funds Confirmation", nickname = "createFundsConfirmation", notes = "", response = OBFundsConfirmationResponse1.class, authorizations = {
-            @Authorization(value = "TPPOAuth2Security", scopes = {
-                    @AuthorizationScope(scope = "fundsconfirmations", description = "Funds Confirmations scope")
-            })
-    })
-    @ApiResponses(value = {
-            @ApiResponse(code = 201, message = "Funds Confirmation Created", response = OBFundsConfirmationResponse1.class),
-            @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
-            @ApiResponse(code = 401, message = "Unauthorized"),
-            @ApiResponse(code = 403, message = "Forbidden"),
-            @ApiResponse(code = 405, message = "Method Not Allowed"),
-            @ApiResponse(code = 406, message = "Not Acceptable"),
-            @ApiResponse(code = 415, message = "Unsupported Media Type"),
-            @ApiResponse(code = 429, message = "Too Many Requests"),
-            @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)})
-
-    @RequestMapping(value = FUNDS_CONFIRMATION_PATH,
-            produces = {"application/json; charset=utf-8"},
-            consumes = {"application/json; charset=utf-8"},
-            method = RequestMethod.POST)
-    ResponseEntity<OBFundsConfirmationResponse1> createFundsConfirmation(
-            @ApiParam(value = "Default", required = true)
-            @Valid
-            @RequestBody OBFundsConfirmation1 obFundsConfirmation1,
-
-            @ApiParam(value = "An Authorisation Token as per https://tools.ietf.org/html/rfc6750", required = true)
-            @RequestHeader(value = "Authorization", required = true) String authorization,
-
-            @ApiParam(value = "The time when the PSU last logged in with the TPP.  All dates in the HTTP headers are represented as RFC 7231 Full Dates. An example is below:  Sun, 10 Sep 2017 19:43:31 UTC")
-            @RequestHeader(value = "x-fapi-auth-date", required = false)
-            @DateTimeFormat(pattern = ApiConstants.HTTP_DATE_FORMAT) DateTime xFapiAuthDate,
-
-            @ApiParam(value = "The PSU's IP address if the PSU is currently logged in with the TPP.")
-            @RequestHeader(value = "x-fapi-customer-ip-address", required = false) String xFapiCustomerIpAddress,
-
-            @ApiParam(value = "An RFC4122 UID used as a correlation id.")
-            @RequestHeader(value = "x-fapi-interaction-id", required = false) String xFapiInteractionId,
-
-            @ApiParam(value = "Indicates the user-agent that the PSU is using.")
-            @RequestHeader(value = "x-customer-user-agent", required = false) String xCustomerUserAgent,
-
-            @ApiParam(value = "OAuth2.0 client_id of the ApiClient making the request")
-            @RequestHeader(value = "x-api-client-id") String apiClientId,
-
-            HttpServletRequest request,
-
-            Principal principal
-    ) throws OBErrorResponseException;
-}
